feat :Creat CRUD Sellers

+++ backend/src/models/purchasesModel.go
@@ -4,7 +4,7 @@ import "github.com/lib/pq"
 
 type Purchase struct {
 	Id        string        `json:"ID" gorm:"primary_key"`
-	Price     float64       `json:"price"`
+	Price     float32       `json:"price"`
 	Products  pq.Int64Array `gorm:"type:integer[]"`
 	Custurmer string        `json:"custumer"`
 }
